modified for depccg and issue#20

+++ en/similarity_en_mp_any.sh
@@ -132,7 +132,7 @@ function parse_easyccg() {
     -model ${candc_dir}/models/ner \
     -ofmt "%w|%p|%n \n" \
     2>/dev/null | \
-  java -jar ${easyccg_dir}/easyccg.jar \
+  /usr/apps.sp3/nosupport/java/jre/bin/java -jar ${easyccg_dir}/easyccg.jar \
     --model ${easyccg_dir}/model \
     -i POSandNERtagged \
     -o extended \
@@ -149,17 +149,21 @@ function parse_easyccg() {
 function select_answer() {
   answer1_fname=$1
   answer2_fname=$2
+  answer3_fname=$3
   base_fname1=`echo ${answer1_fname##*/} | sed 's/.answer//g'`
   base_fname2=`echo ${answer2_fname##*/} | sed 's/.answer//g'`
+  base_fname3=`echo ${answer3_fname##*/} | sed 's/.answer//g'`
   sentences_basename=${base_fname1/.candc/}
-  if [ ! -e $answer1_fname ] && [ ! -e $answer2_fname ]; then
+  if [ ! -e $answer1_fname ] && [ ! -e $answer2_fname ] && [ ! -e $answer3_fname ]; then
     echo "unknown" > ${answer1_fname/.candc/};
     prediction_fname=""
-  elif [ ! -e $answer1_fname ]; then
+  elif [ ! -e $answer1_fname ] && [ ! -e $answer3_fname ]; then
     prediction_fname=$base_fname2
-  elif [ ! -e $answer2_fname ]; then
+  elif [ ! -e $answer2_fname ] && [ ! -e $answer3_fname ]; then
     prediction_fname=$base_fname1
-  elif [ -e $answer1_fname ] && [ -e $answer2_fname ]; then
+  elif [ ! -e $answer1_fname ] && [ ! -e $answer2_fname ]; then
+    prediction_fname=$base_fname3
+  elif [ -e $answer1_fname ] && [ -e $answer2_fname ] && [ -e $answer3_fname ]; then
     gold_file=${answer1_fname/results/plain}
     gold_answer=`cat ${gold_file/.candc/}` #gold
     if [ "$gold_answer" == "yes" ]; then
@@ -173,41 +177,59 @@ function select_answer() {
     answer1=${answer1/\[}
     answer2=`cat ${answer2_fname}|tr -d '\r'|awk -F , 'NR == 1 {print $1}'` #easyccg
     answer2=${answer2/\[}
+    answer3=`cat ${answer3_fname}|tr -d '\r'|awk -F , 'NR == 1 {print $1}'` #depccg
+    answer3=${answer3/\[}
 
-    #select candc or easyccg
-    # use candc
-    if [ "$answer1" == "1" ] && [ "$answer2" == "0" ]; then
-      prediction_fname=$base_fname1
-    elif [ "$answer1" == "1" ] && [ "$answer2" == "coq_error" ]; then
-      prediction_fname=$base_fname1
-    elif [ "$answer1" == "1" ] && [ "$answer2" == "" ]; then
-      prediction_fname=$base_fname1
-    elif [ "$answer1" == "1" ] && [ "$answer2" == "unknown" ]; then
-      prediction_fname=$base_fname1
-    elif [ "$answer1" == "0.5" ] && [ "$answer2" == "0" ]; then
-      prediction_fname=$base_fname1
-    elif [ "$answer1" == "0.5" ] && [ "$answer2" == "coq_error" ]; then
-      prediction_fname=$base_fname1
-    elif [ "$answer1" == "0.5" ] && [ "$answer2" == "" ]; then
-      prediction_fname=$base_fname1
-    elif [ "$answer1" == "0.5" ] && [ "$answer2" == "unknown" ]; then
-      prediction_fname=$base_fname1
-    elif [ "$answer1" == "0" ] && [ "$answer2" == "coq_error" ]; then
-      prediction_fname=$base_fname1
-    elif [ "$answer1" == "0" ] && [ "$answer2" == "" ]; then
-      prediction_fname=$base_fname1
-    elif [ "$answer1" == "0" ] && [ "$answer2" == "unknown" ]; then
-      prediction_fname=$base_fname1
-    else
-      #use easyccg
-      prediction_fname=$base_fname2
+    #select candc, easyccg or depccg
+    #accuracy: candc > depccg > easyccg
+    answer_level1=("1" "0.5")
+    answer_level2=("0" "coq_error" "unknown" "")
+    answer_level3=("coq_error" "unknown" "")
+    if `echo ${answer_level1[@]} | grep -q "$answer2"` && `echo ${answer_level2[@]} | grep -q "$answer1"` && `echo ${answer_level2[@]} | grep -q "$answer3"`; then
+        prediction_fname=$base_fname2 #easyccg
+    elif [ "$answer2" == "0" ] && `echo ${answer_level3[@]} | grep -q "$answer1"` && `echo ${answer_level3[@]} | grep -q "$answer2"`; then
+        prediction_fname=$base_fname2 #easyccg
+    elif `echo ${answer_level1[@]} | grep -q "$answer3"` && `echo ${answer_level2[@]} | grep -q "$answer1"`; then
+        prediction_fname=$base_fname3 #depccg
+    elif [ "$answer1" == "0" ] && `echo ${answer_level3[@]} | grep -q "$answer1"`; then
+        prediction_fname=$base_fname3 #depccg
+    else:
+        prediction_fname=$base_fname1 #candc
     fi
+    #if [ "$answer1" == "1" ] && [ "$answer2" == "0" ]; then
+    #  prediction_fname=$base_fname1
+    #elif [ "$answer1" == "1" ] && [ "$answer2" == "coq_error" ]; then
+    #  prediction_fname=$base_fname1
+    #elif [ "$answer1" == "1" ] && [ "$answer2" == "" ]; then
+    #  prediction_fname=$base_fname1
+    #elif [ "$answer1" == "1" ] && [ "$answer2" == "unknown" ]; then
+    #  prediction_fname=$base_fname1
+    #elif [ "$answer1" == "0.5" ] && [ "$answer2" == "0" ]; then
+    #  prediction_fname=$base_fname1
+    #elif [ "$answer1" == "0.5" ] && [ "$answer2" == "coq_error" ]; then
+    #  prediction_fname=$base_fname1
+    #elif [ "$answer1" == "0.5" ] && [ "$answer2" == "" ]; then
+    #  prediction_fname=$base_fname1
+    #elif [ "$answer1" == "0.5" ] && [ "$answer2" == "unknown" ]; then
+    #  prediction_fname=$base_fname1
+    #elif [ "$answer1" == "0" ] && [ "$answer2" == "coq_error" ]; then
+    #  prediction_fname=$base_fname1
+    #elif [ "$answer1" == "0" ] && [ "$answer2" == "" ]; then
+    #  prediction_fname=$base_fname1
+    #elif [ "$answer1" == "0" ] && [ "$answer2" == "unknown" ]; then
+    #  prediction_fname=$base_fname1
+    #else
+    #  #use easyccg
+    #  prediction_fname=$base_fname2
+    #fi
     #if there is gold answer, check gold answer
     if [ -n "$gold_answer" ]; then
       if [ "$gold_answer" == "$answer1" ]; then
         prediction_fname=$base_fname1 #candc
       elif [ "$gold_answer" == "$answer2" ]; then
         prediction_fname=$base_fname2 #easyccg
+      elif [ "$gold_answer" == "$answer3" ]; then
+        prediction_fname=$base_fname3 #depccg
       fi
     fi
   fi
@@ -246,8 +268,9 @@ fi
 # Judge entailment with a theorem prover (Coq, at the moment).
 if [ ! -e "${results_dir}/${sentences_basename/.tok/.answer}" ]; then
   start_time=`python -c 'import time; print(time.time())'`
-  for parser in `cat en/parser_location.txt`; do
-    parser_name=`echo $parser | awk -F':' '{print $1}'`
+  #for parser in `cat parser_location.txt`; do
+  #  parser_name=`echo $parser | awk -F':' '{print $1}'`
+  for parser_name in {candc,easyccg,depccg}; do
     if [ ! -e "${results_dir}/${sentences_basename}.${parser_name}.answer" ]; then
       if [ "$word2vec" == "word2vec" ]; then
         timeout 600 python scripts/prove_w2v.py \
@@ -276,6 +299,7 @@ if [ ! -e "${results_dir}/${sentences_basename/.tok/.answer}" ]; then
   echo $proving_time > ${results_dir}/${sentences_basename}.time
   select_answer \
     ${results_dir}/${sentences_basename}.candc.answer \
-    ${results_dir}/${sentences_basename}.easyccg.answer
+    ${results_dir}/${sentences_basename}.easyccg.answer \
+    ${results_dir}/${sentences_basename}.depccg.answer
 fi
 echo "Judged entailment for $parsed_dir/${sentences_basename}.sem.xml "`cat ${results_dir}/${sentences_basename}.answer`
